reorganize files, added colors file
