+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { RouteHandler } from 'react-router';
-import ContactsKeeperHeader from './Pages/Components/ContactsKeeperHeader.component.js'
-import ContactsKeeperFooter from './Pages/Components/ContactsKeeperFooter.component.js'
+import ContactsKeeperHeader from './Pages/ContactsKeeperPage/Components/ContactsKeeperHeader.component.js'
+import ContactsKeeperFooter from './Pages/ContactsKeeperPage/Components/ContactsKeeperFooter.component.js'
 import {fullHeight, appContainer} from './App.scss'
 export default class App extends Component {
     constructor() {
