+++ src/Components/Table/Table.scss
@@ -1,8 +1,9 @@
-@import "../../scss/flex";
+@import "../../scss/flex",
+"../../scss/colors";
 
 .table {
   @extend .flexColumn;
-  border: 1px solid #676767;
+  border: 1px solid $input-border-grey;
   margin: 20px 0;
   border-radius: 3px;
 }
