+++ src/Components/Table/TableCell.scss
@@ -10,8 +10,8 @@
 .tableCellRow {
   @extend .tableCell;
   padding: 5px 0;
-  border-right: 1px solid #E7E7E7;
-  border-top: 1px solid #E7E7E7;
+  border-right: 1px solid $table-cell-border-grey;
+  border-top: 1px solid $table-cell-border-grey;
 }
 
 .notesRowCell {
