+++ src/scss/colors.scss
@@ -1,0 +1,5 @@
+$procore-blue: #216FB5;
+$procore-orange: #f58025;
+$modal-input-grey: #888888;
+$input-border-grey: #676767;
+$table-cell-border-grey: #E7E7E7;
